Move some files from JetpackStatsWidgets to JetpackStatsWidgetsCore
Notice that JetpackIntents now links JetpackStatsWidgetsCore.
This was necessary to compile the target, even though I couldn't see any
of the types from JetpackStatsWidgetsCore being used in the Intents
extension source.

Also notice that we had to link JetpackStatsWidgetsCore to WordPress and
Jetpack, which kinda defeats the purpose for which we created the
framework in the first place, doesn't it?

I am hopeful that:

1. With some refactor, we might be able to remove the knowledge of this
   framework from the apps; alternatively
2. We might extract an additional "shared package" with the types and
   protocols that apps and widgets use
3. Regardless, it should be possible to not link it in WordPress, given
   that Jetpack is the only app that has the stats widget.
