+++ Modules/Sources/JetpackStatsWidgetsCore/WidgetDataCacheReader.swift
@@ -1,0 +1,4 @@
+public protocol WidgetDataCacheReader {
+    func widgetData<T: HomeWidgetData>(for siteID: String) -> T?
+    func widgetData<T: HomeWidgetData>() -> [T]?
+}
