+++ WordPress/JetpackStatsWidgets/Helpers/WidgetDataReader.swift
@@ -1,9 +1,5 @@
 import Foundation
-
-protocol WidgetDataCacheReader {
-    func widgetData<T: HomeWidgetData>(for siteID: String) -> T?
-    func widgetData<T: HomeWidgetData>() -> [T]?
-}
+import JetpackStatsWidgetsCore
 
 enum WidgetDataReadError: Error {
     case jetpackFeatureDisabled
