+++ WordPress/JetpackStatsWidgets/LockScreenWidgets/Configs/LockScreenTodayLikesCommentsStatWidgetConfig.swift
@@ -1,4 +1,5 @@
 import WidgetKit
+import JetpackStatsWidgetsCore
 
 @available(iOS 16.0, *)
 struct LockScreenTodayLikesCommentsStatWidgetConfig: LockScreenStatsWidgetConfig {
