+++ WordPress/JetpackStatsWidgets/LockScreenWidgets/LockScreenStatsWidget.swift
@@ -1,5 +1,6 @@
 import WidgetKit
 import SwiftUI
+import JetpackStatsWidgetsCore
 
 @available(iOS 16.0, *)
 struct LockScreenStatsWidget<T: LockScreenStatsWidgetConfig>: Widget {
