+++ WordPress/JetpackStatsWidgets/LockScreenWidgets/ViewProvider/LockScreenMultiStatWidgetViewProvider.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import JetpackStatsWidgetsCore
 
 @available(iOS 16.0, *)
 struct LockScreenMultiStatWidgetViewProvider<WidgetData: HomeWidgetData>: LockScreenStatsWidgetsViewProvider {
