+++ WordPress/JetpackStatsWidgets/LockScreenWidgets/ViewProvider/LockScreenSingleStatWidgetViewProvider.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import JetpackStatsWidgetsCore
 
 @available(iOS 16.0, *)
 struct LockScreenSingleStatWidgetViewProvider<WidgetData: HomeWidgetData>: LockScreenStatsWidgetsViewProvider {
