+++ WordPress/JetpackStatsWidgets/LockScreenWidgets/Views/LockScreenStatsWidgetsView.swift
@@ -1,5 +1,6 @@
 import SwiftUI
 import WidgetKit
+import JetpackStatsWidgetsCore
 
 protocol LockScreenStatsWidgetsViewProvider<Data> {
     associatedtype SiteSelectedView: View
