+++ WordPress/JetpackStatsWidgets/Model/HomeWidgetAllTimeData.swift
@@ -1,3 +1,4 @@
+import JetpackStatsWidgetsCore
 
 struct HomeWidgetAllTimeData: HomeWidgetData {
