+++ WordPress/JetpackStatsWidgets/Model/HomeWidgetData.swift
@@ -1,19 +1,7 @@
-import WidgetKit
-
-protocol HomeWidgetData: Codable {
-
-    var siteID: Int { get }
-    var siteName: String { get }
-    var url: String { get }
-    var timeZone: TimeZone { get }
-    var date: Date { get }
-    var statsURL: URL? { get }
-
-    static var filename: String { get }
-}
-
+import JetpackStatsWidgetsCore
 
 // MARK: - Local cache
+
 extension HomeWidgetData {
 
     static func read(from cache: HomeWidgetCache<Self>? = nil) -> [Int: Self]? {
