+++ WordPress/WordPressTest/Widgets/WidgetDataReaderTests.swift
@@ -1,3 +1,4 @@
+import JetpackStatsWidgetsCore
 import XCTest
 @testable import WordPress
